Fix regression in Coqide's "forward one command" command
In Coqide 8.5pl2, "forward one command" (down arrow) always repositions the insertion point at the end of the phrase being executed, just after the final ".".

In Coqide 8.4, the insertion point is not moved if it is after the end of the executed phrase.  The insertion point is moved only if it falls behind the phrase being executed.

I find the 8.5 behavior to be a regression over the lot more useful 8.4 behavior.  This commit restores the 8.4 behavior of "forward one command".