ENH: Termops.compact_named_context_reverse
I have noticed that CoqIDE compacts hypotheses in some cases but not in other cases.

E.g. in this case:

  (* artificial goal *)
  Goal forall (P:Prop) (Q:Prop) (n:nat) (m:nat), P -> P.
    intros.

hypotheses are compacted. That is the context (as displayed by CoqIDE) looks this way:

  P, Q : Prop
  n, m : nat
  H : P

Which is nice.

But, in this case:

  Goal forall (P:Prop) (n:nat) (Q:Prop) (m:nat), P -> P.
    intros.

hypotheses are not compacted. CoqIDE displays them like this:

  P : Prop
  n : nat
  Q : Prop
  m : nat
  H : P

As if we wanted to compact hypotheses but manage to do that only in some, not all cases.

This commit changes the behavior so that CoqIDE displays compacted hypotheses
whenever that compaction is possible.

updating the test-cases to reflect changes introduced by commit "ab4a929"