Fix incorrect glob data for module symbols (bug #2336).
The logic was backward: if the path of a symbol was a prefix of the
current path, then the current path (without sections) was used. But what
we want is that, if the current path (without sections) is a prefix of the
path of a symbol, then the former should be used.

This fixes about 1,600 broken links in the documentation of the standard
library.