[pp] Fix bug 4842 (Time prints in multiple lines)
This commit proposes a fix for
https://coq.inria.fr/bugs/show_bug.cgi?id=4842

Here, we face a bit of a dilemma: we need to choose between sending the
time report to the user in one or two messages.

In the one message approach, we send the whole report once the command
has been completed. In the two messages approach, we send one message at
the start and one at completion. However, in the second approach, we
lose some control of the formatting.

I've opted to send just one message for timing event: I think it is a
better choice for the IDEs, and it is the easiest path to preserve
output compatibility this late in the release process.

However, note that having a finer control over messages layout could
make sense, but IMVHO in order to achieve a satisfactory implementation
we would need to implement `pp_stdcmds` serialization first.