Avoid canonizing the same paths over and over.
The number of path canonizations was quadratic in the number of potential
plugin directories. This patch makes it linear; on a standard Coq tree,
this brings the number of chdir and getcwd system calls from more than
1,000 down to about 200 at startup.

This also fixes a bug where the Cd vernacular command could cause plugins
to break since relative paths were used to locate them.