Warning about similar notations now up to alpha-conversion.
This allows to define on purpose the very same notation in different
files, as currently the notations for *, +, - in Nat.v and Peano.v
(with the first one using variables n and m and the second one using
the default variables used by Infix, namely x and y).

This makes also the "notation-overridden" warning less enigmatic
facing two notations which are the same up to the choice of names.