STM: compute the correct state for edited Qed (#5086)
When a proof is 're-opened', the Qed node does not change.
Still the STM has to install the old state (where only
the future proof has to be updated).  This bit was missing.

Why was it working: the bug happens only if you
reopen the very last proof, i.e. there is no
sentence that stays valid after the Qed.  If there
is such a sentence, its state was computed correctly
before, and is not changed.  If it is the very last,
then the next state is based on the wrong one...