Set the default LtacProf cutoff to 2%
This was the original value from Tobias' code.  When a user passes
-profile-ltac on the command line, or inserts [Show Ltac Profile] in the
document, the most useful default behavior is to not overload them with
useless information.  When GUI clients want to display fancier profiling
information, there is no cost to the user to requiring them to specify
what cutoff they want.  If the GUI client does not have any special
LtacProf handling, the most useful presentation is again the one that
cuts off the display at 2% total time.