Move bullet detection from lexer to parser (bug #5102).
That way, bullet detection no longer depends on a global variable
indicating whether a line is starting. This causes a small change in the
recognized language. Before the commit, "--++" was recognized as a bullet
"--" followed by a keyword "++" when at the start of a line; now it is
always recognized as a keyword "--++".

This also fixes a bug in Tok.to_string as a side-effect.