Fix the semantics of the import of options.
Before this patch, Set and Local Set behaved the same outside of a section,
i.e. they disappeared at module closing. Now, setting an option without a
qualifier allows to activate it whenever the module it lives in is imported.
This is a much saner behaviour that allows to scope options properly.

Fixing the Coq source files was a mere matter of systematically replacing
every Set / Unset invokations by a corresponding Local Set / Unset command.