Fix #4959: Ltac match fails to match Type with Type
Non-linear Ltac pattern matching can now produce universe constraints.
Previously, it was comparing universes up to syntactic equality, which
is too restrictive.

However, this fix removes an optimization in tacting_matching,
where functions taking an env and an evar_map were specialized using
a first class functor, instead of threading explicitely data that did
not change. This fix requires the evar_map to change when evaluating the
pattern matching, so I had to remove it from the functor, and we are now
creating more closures.