[pp] Try to properly tag error messages in cError.
In order to get proper coloring, we must tag the headers of error
messages in `CError`.

This should fix bug
https://coq.inria.fr/bugs/show_bug.cgi?id=5135

However, note that this could interact badly with the richpp printing
used by the IDE. At this level, we have no clue which tag we'd like to
apply, as we know (and shouldn't) nothing about the top level backend.

Thus, for now I've selected the console printer, hoping that the
`Richpp` won't crash the IDE.