Quick fix to unification regression #4955.
The commit which caused the regression (5ea2539d4a) looks reasonable.
However, it happens that this commit made that unification started in
the #4955 example to follow a path with problems of the form
"proj ?x == ?x" which clearly are unsolvable (both ?x have the same
instance).

We hack the corresponding very permissive occur-check function so that
it is a bit less permissive.

One day, this occur-check function would have to be rewritten in a
more stricter way.

----------------------------------------------------------------------
Extra comments:

I could list several functions for occur check of evars.

Four of them are too strict in the sense that they do not take into
account occurrences of evars which may disappear by reduction, nor
evars which have instances made in such a way that the occur-check is
solvable (as in "fst ?x[y:=(0,0)] = ?x[y:=0]"). These are:
- Termops.occur_evar for clenv, evar_refiner, tactic unification
  - syntactic check without any normalization, even on defined evars
- Evarutil.occur_evar_upto for refine and the V82 compatibility mode
  - syntactic check without any normalization but inlining of defined evars
- Evarsolve.occur_evar_upto_types for evar_define
  - syntactic check without any normalization but inlining of defined evars
  - occur-check in the type of evars met
  - optimization for not visiting several time the same evar
- Evarsolve.noccur_evar for pattern unification and for inversion of
  substitution (evar/evar or evar/term problems)
  - syntactic check without any normalization but inlining of defined evars
  - occur-check in the type of evars met in arguments of projections
  - occur-check in the type of variables met in arguments of projections
  - optimization for not visiting several time the same evar
  - optimization for not visiting several time the type of the same variable
  - note: to go this way, it seems to me that it should check also in
    all types which are a cut-formula in the expression

One is much too lax:
- Evarconv.occur_rigidly
  - no recursive check except on the types of "forall" and sometimes
    in the arguments of an application
  - note: there is obviously a large room for refinements without
    loosing solutions