Removing output test for module qualification.
I do not know how to provide a proper test in 8.5, as the location on my
machine appears in the error printed when loading the file. Adding a Fail
on the End command does not help much either, because it simply does not
print anything.

Do not merge this commit in 8.6, we still want a test there.