Removing an horribly designed syntax for eauto bfs.
Before this commit, it was possible to use eauto 4 3 to call
eauto in breadth-first search with depth 3 (4 being
a completely dummy argument).
This horrible syntax does not seem to be documented anywhere,
not even in CHANGES so there is nothing against removing it.
It is not used in the stdlib.

NB: this syntax has existed at least since 2002 when .ml4
files were introduced.