Introduce new tactic dfs eauto.
dfs eauto is a synonym for try typeclasses eauto with core
without typeclass_instances, with bounded search-depth
and with ~only_classes:false.
Thus it tries to be as close as possible to the current eauto
while being based on a new implementation.
The optional depth argument has a slightly different semantics
(the proof search reaches its limit earlier).
To write forward compatible code in 8.6 you can use dfs eauto
since the current implementation of eauto is going to get
removed eventually.