Adding the possibility to give a level to generic argument printers.
This allows to call the tactic printer with the correct printer when
printing Tactic Notation or TACTIC EXTEND rules.

However, this is not very cleanly done, since printers who don't have
a level now takes as argument a dummy level. One would need to use the
entry type to know if a level is needed or not.