A partial fix to #5220 ("destruct"'s section variable test too weak).
To detect that a variable is (maybe) coming from the section context,
we check the type instead of the only the name.

Of course, this is not enough in the absolute since it is still
possible to have a section variable and a goal variable of same name,
but this is a little progress while a proper detection of section
variables is done.