Fix #5183 - Two CoqIDE crash errors
When opening a file without extension, an uncaught exception was
occurring.

Note that this fix is not complete, since the "Compile Buffer" command
still fails. This is because of a limitation of coqc which appends the
".v" extension to its argument even if it already existed (and even if
it doesn't exist with the extension!).