Avoid concurrent runs when producing html documentation (bug #5269).
Make does not allow for rules that produce multiple outputs (unless they
are pattern rules). As a consequence, the hacha rule could be run several
times concurrently, thus causing doc/refman/html to be deleted under the
feet of some runs.

This commit fixes the issue by turning the rule into a single-output rule
and adding a dummy rule to handle all the other indexes. Note that this is
not a perfect solution since, if the user were to manually delete one of
the auxiliary index, it would not be rebuilt until the main index is.