[stm] Reenable Show Script command.
We extend `Stm.vernac_interp` so it can handle the commands it should
by level. This reenables `Show Script` handling, and this
interpretation function should handle more commands in the future such
as Load.

However note that we must first refactor the parsing state handling a
bit and remove the legacy `Stm.interp` before that.