Fix SR breakage due to allowing fixpoints on non-rec values
We limit fixpoints to Finite inductive types, so that BiFinite
inductives (non-recursive records) are excluded from fixpoint
construction. This is a regression in the sense that e.g. fixpoints
on unit records were allowed before. Primitive records with
eta-conversion are included in the BiFinite types.

Fix deprecation

Fix error message, the inductive type needs to be recursive for fix to work