Relaxing syntax of parameters in well-formedness of inductive types.
We support that parameters are considered up to conversion rather than
syntactically being variables in the definition of an inductive type.

This allows for instance to support the following:

Inductive sig {A:Type} (P:A -> Prop) : Type :=
    exist : forall x:A, P x -> { x : A | P x }
where "{ x : A |  P }" := (sig (fun x : A => P)) : type_scope.

where the parameter P appears not as a variable but as an eta-expanded
variable.