Fixing a little bug in using the "where" clause for inductive types.
This was not working when the inductive type had implicit parameters.

This could still be improved. E.g. the following still does not work:

Reserved Notation "#".
Inductive I {A:Type} := C : # where "#" := I.

But it should be made working by taking care of adding the mandatory
implicit argument A at the time # is interpreted as [@I _]
(i.e., technically speaking, using expl_impls in interp_notation).